[android] Fix google drive upload
