+++ app_manager/appmanager/src/main/java/app/editors/manager/storages/googledrive/ui/fragments/operations/DocsGoogleDriveOperationFragment.kt
@@ -1,16 +1,18 @@
 package app.editors.manager.storages.googledrive.ui.fragments.operations
 
+import android.net.Uri
 import app.documents.core.network.ApiContract
 import app.editors.manager.BuildConfig
 import app.editors.manager.app.App
 import app.editors.manager.mvp.models.account.Storage
 import app.editors.manager.storages.base.fragment.BaseStorageOperationsFragment
+import app.editors.manager.storages.base.view.DocsGoogleDriveView
 import app.editors.manager.storages.googledrive.mvp.presenters.DocsGoogleDrivePresenter
 import app.editors.manager.storages.googledrive.ui.fragments.GoogleDriveSignInFragment
 import app.editors.manager.ui.fragments.operations.DocsCloudOperationFragment
 import moxy.presenter.InjectPresenter
 
-class DocsGoogleDriveOperationFragment: BaseStorageOperationsFragment() {
+class DocsGoogleDriveOperationFragment: BaseStorageOperationsFragment(), DocsGoogleDriveView {
 
     companion object {
 
@@ -23,6 +25,11 @@ class DocsGoogleDriveOperationFragment: BaseStorageOperationsFragment() {
     override lateinit var presenter: DocsGoogleDrivePresenter
 
     override fun getOperationsPresenter() = presenter
+
+    override fun onUpload(uploadUris: List<Uri>, folderId: String, fileId: String, tag: String) {
+        // Stub
+    }
+
     override fun onRefreshToken() {
         val storage = Storage(
             ApiContract.Storage.GOOGLEDRIVE,
